backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.2
python-multipart==0.0.9
authlib==1.3.2
itsdangerous==2.2.0
apscheduler==3.10.4
redis==5.1.1
celery==5.4.0
python-dotenv==1.0.1
